chore: integrated local whisper + restructuring
