+++ .gitignore
@@ -1,3 +1,5 @@
+ggml-*.bin
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
@@ -7,7 +9,7 @@ __pycache__/
 *.so
 
 # Distribution / packaging
-.Python
+.Pythongit
 build/
 develop-eggs/
 dist/
