+++ OS/01/conversations/user.json
@@ -1,1 +1,1 @@
-[{"role": "user", "type": "message", "content": "This is a microphone. I also have an extra microphone.\n"}]
+[{"role": "user", "type": "message", "content": " Hello, how you doing?\n"}]
