+++ OS/01/local_stt/whisper-rust/.gitignore
