+++ OS/01/local_stt/whisper-rust/src/transcribe.rs
