--- OS/_archive/core/stt/whisper-rust/.env.example
@@ -1,1 +1,0 @@
-WHISPER_MODEL_PATH=/path/to/ggml-tiny.en.bin
